+++ saois/commands/__init__.py
@@ -1,0 +1,4 @@
+# SAOIS Commands Module
+from .main import run_command
+
+__all__ = ['run_command']